make blueprint state into single class.
