+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/ILyricEditorState.cs
@@ -2,14 +2,11 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using osu.Framework.Bindables;
-using osu.Game.Rulesets.Karaoke.Objects;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
 {
     public interface ILyricEditorState
     {
-        #region general
-
         Bindable<LyricEditorMode> BindableMode { get; }
 
         LyricEditorMode Mode { get; set; }
@@ -19,21 +16,5 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
         BindableBool BindableAutoFocusEditLyric { get; }
 
         BindableInt BindableAutoFocusEditLyricSkipRows { get; }
-
-        #endregion
-
-        # region blueprint
-
-        BindableList<TimeTag> SelectedTimeTags { get; }
-
-        BindableList<RubyTag> SelectedRubyTags { get; }
-
-        BindableList<RomajiTag> SelectedRomajiTags { get; }
-
-        void ClearSelectedTimeTags();
-
-        void ClearSelectedTextTags();
-
-        #endregion
     }
 }
