--- osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricEditor_State.cs
@@ -1,19 +1,0 @@
-﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
-{
-    public partial class LyricEditor
-    {
-        public void ClearSelectedTimeTags()
-        {
-            SelectedTimeTags.Clear();
-        }
-
-        public void ClearSelectedTextTags()
-        {
-            SelectedRubyTags.Clear();
-            SelectedRomajiTags.Clear();
-        }
-    }
-}
